third_party/WebKit/LayoutTests/inspector/tracing/timeline-js-callstacks.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../tracing-test.js"></script>
<script>
function test()
{
    var sessionId = "6.23";
    var rawTraceEvents = [
    {
        "args": {
          "name": "Renderer"
        },
        "cat": "__metadata",
        "name": "process_name",
        "ph": "M",
        "pid": 17851,
        "tid": 23,
        "ts": 0
    },
    {
        "args": {
            "name": "CrRendererMain"
        },
        "cat": "__metadata",
        "name": "thread_name",
        "ph": "M",
        "pid": 17851,
        "tid": 23,
        "ts": 0
    },
    {
        "args": {
            "sessionId": sessionId
        },
        "cat": "disabled-by-default-devtools.timeline",
        "name": "TracingStartedInPage",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 100000,
        "tts": 606543
    },
    {
        "args": {},
        "cat": "disabled-by-default-devtools.timeline",
        "name": "Program",
        "ph": "B",
        "pid": 17851,
        "tid": 23,
        "ts": 200000,
        "tts": 5612442
    },
    {
        "args": {},
        "cat": "disabled-by-default-devtools.timeline",
        "name": "JSSample",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 208000,
        "tts": 1758056
    },
    {
        "args": {
            "stack": [
                { "functionName": "c", "scriptId": 1 },
                { "functionName": "b", "scriptId": 1 },
                { "functionName": "a", "scriptId": 1 }
            ]
        },
        "cat": "disabled-by-default-devtools.timeline",
        "name": "CallStack",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 208000,
        "tts": 1758056
    },
    {
        "args": {},
        "cat": "disabled-by-default-devtools.timeline",
        "name": "FunctionCall",
        "ph": "X",
        "pid": 17851,
        "tid": 23,
        "ts": 210000,
        "dur": 30000,
        "tts": 5612442
    },
    {
        "args": {},
        "cat": "disabled-by-default-devtools.timeline",
        "name": "JSSample",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 211000,
        "tts": 1758056
    },
    {
        "args": {
            "stack": [
                { "functionName": "c", "scriptId": 1 },
                { "functionName": "b", "scriptId": 1 },
                { "functionName": "a", "scriptId": 1 }
            ]
        },
        "cat": "disabled-by-default-devtools.timeline",
        "name": "CallStack",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 211000,
        "tts": 1758056
    },
    {
        "args": {},
        "cat": "disabled-by-default-devtools.timeline",
        "name": "JSSample",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 219875,
        "tts": 1758056
    },
    {
        "args": {
            "stack": [
                { "functionName": "c", "scriptId": 1 },
                { "functionName": "b", "scriptId": 1 },
                { "functionName": "a", "scriptId": 1 }
            ]
        },
        "cat": "disabled-by-default-devtools.timeline",
        "name": "CallStack",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 219875,
        "tts": 1758056
    },
    {
        "args": {
            "frame": "0x2f7b63884000"
        },
        "cat": "disabled-by-default-devtools.timeline",
        "name": "InvalidateLayout",
        "ph": "X",
        "pid": 17851,
        "tid": 23,
        "ts": 220000,
        "dur": 7000,
        "tts": 1758056
    },
    {
        "args": {
            "stack": [
                { "functionName": "b", "scriptId": 1 },
                { "functionName": "a", "scriptId": 1 }
            ]
        },
        "cat": "disabled-by-default-devtools.timeline",
        "name": "CallStack",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 220000,
        "tts": 1758056
    },
    {
        "args": {},
        "cat": "disabled-by-default-devtools.timeline",
        "name": "JSSample",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 220125,
        "tts": 1758056
    },
    {
        "args": {
            "stack": [
                { "functionName": "c", "scriptId": 1 },
                { "functionName": "b", "scriptId": 1 },
                { "functionName": "a", "scriptId": 1 }
            ]
        },
        "cat": "disabled-by-default-devtools.timeline",
        "name": "CallStack",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 220125,
        "tts": 1758056
    },
    {
        "args": {
            "frame": "0x2f7b63884000"
        },
        "cat": "disabled-by-default-devtools.timeline",
        "name": "InvalidateLayout",
        "ph": "X",
        "pid": 17851,
        "tid": 23,
        "ts": 221000,
        "dur": 3000,
        "tts": 1758056
    },
    {
        "args": {
            "stack": [
                { "functionName": "e", "scriptId": 1 },
                { "functionName": "d", "scriptId": 1 },
                { "functionName": "b", "scriptId": 1 },
                { "functionName": "a", "scriptId": 1 }
            ]
        },
        "cat": "disabled-by-default-devtools.timeline",
        "name": "CallStack",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 221000,
        "tts": 1758056
    },
    {
        "args": {},
        "cat": "disabled-by-default-devtools.timeline",
        "name": "JSSample",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 222000,
        "tts": 1758056
    },
    {
        "args": {
            "stack": [
                { "functionName": "g", "scriptId": 1 },
                { "functionName": "f", "scriptId": 1 },
                { "functionName": "b", "scriptId": 1 },
                { "functionName": "a", "scriptId": 1 }
            ]
        },
        "cat": "disabled-by-default-devtools.timeline",
        "name": "CallStack",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 222000,
        "tts": 1758056
    },
    {
        "args": {},
        "cat": "disabled-by-default-devtools.timeline",
        "name": "JSSample",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 227125,
        "tts": 1758056
    },
    {
        "args": {
            "stack": [
                { "functionName": "g", "scriptId": 1 },
                { "functionName": "e", "scriptId": 1 },
                { "functionName": "b", "scriptId": 1 },
                { "functionName": "a", "scriptId": 1 }
            ]
        },
        "cat": "disabled-by-default-devtools.timeline",
        "name": "CallStack",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 227125,
        "tts": 1758056
    },
    {
        "args": {},
        "cat": "disabled-by-default-devtools.timeline",
        "name": "JSSample",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 227250,
        "tts": 1758056
    },
    {
        "args": {
            "stack": [
                { "functionName": "g", "scriptId": 1 },
                { "functionName": "e", "scriptId": 1 },
                { "functionName": "b", "scriptId": 1 },
                { "functionName": "a", "scriptId": 1 }
            ]
        },
        "cat": "disabled-by-default-devtools.timeline",
        "name": "CallStack",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 227250,
        "tts": 1758056
    },
    {
        "args": {},
        "cat": "disabled-by-default-devtools.timeline",
        "name": "JSSample",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 230000,
        "tts": 1758056
    },
    {
        "args": {
            "stack": [
                { "functionName": "a", "scriptId": 1 },
                { "functionName": "l", "scriptId": 1 },
                { "functionName": "f", "scriptId": 1 }
            ]
        },
        "cat": "disabled-by-default-devtools.timeline",
        "name": "CallStack",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 230000,
        "tts": 1758056
    },
    {
        "args": {},
        "cat": "disabled-by-default-devtools.timeline",
        "name": "JSSample",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 230125,
        "tts": 1758056
    },
    {
        "args": {
            "stack": [
                { "functionName": "a", "scriptId": 1 },
                { "functionName": "l", "scriptId": 1 },
                { "functionName": "f", "scriptId": 1 }
            ]
        },
        "cat": "disabled-by-default-devtools.timeline",
        "name": "CallStack",
        "ph": "I",
        "pid": 17851,
        "tid": 23,
        "ts": 230125,
        "tts": 1758056
    },
    {
        "args": {},
        "cat": "disabled-by-default-devtools.timeline",
        "name": "Program",
        "ph": "E",
        "pid": 17851,
        "tid": 23,
        "ts": 500000,
        "tts": 5612506
    }
    ];

    var tracingTimelineModel = InspectorTest.tracingTimelineModel();
    tracingTimelineModel.setEventsForTest(rawTraceEvents);
    var dataProvider = new WebInspector.TimelineFlameChartDataProvider(tracingTimelineModel, null);
    var events = dataProvider._generateJSFrameEvents(tracingTimelineModel.inspectedTargetEvents());
    events.map(function(e) {
        return e.name + ": " + e.startTime + "(" + (e.duration || 0) + ") " + e.args.data.functionName;
    }).forEach(InspectorTest.addResult);
    InspectorTest.completeTest();
}

</script>
</head>
<body onload="runTest()">
</body>
</html>
