third_party/WebKit/LayoutTests/inspector/tracing/timeline-js-callstacks-expected.txt

Program: 200.000 / 300.000 
FunctionCall: 210.000 / 30.000 
JSFrame: 211.000 / 0.226 a
JSFrame: 211.000 / 0.226 b
JSFrame: 211.000 / 0.226 c
JSFrame: 219.875 / 8.155 a
JSFrame: 219.875 / 8.155 b
JSFrame: 219.875 / 7.250 c
InvalidateLayout: 220.000 / 7.000 
InvalidateLayout: 221.000 / 3.000 
JSFrame: 227.125 / 0.905 e
JSFrame: 227.125 / 0.905 g
JSFrame: 230.000 / 0.833 f
JSFrame: 230.000 / 0.833 l
JSFrame: 230.000 / 0.833 a
Layout: 230.010 / 0.100 
FunctionCall: 250.000 / 10.000 
FunctionCall: 251.000 / 1.000 
JSFrame: 251.000 / 0.882 x
JSFrame: 251.000 / 0.882 y
JSFrame: 251.000 / 0.200 z
JSFrame: 251.000 / 0.200 w
JSFrame: 251.200 / 0.200 w
JSFrame: 253.000 / 0.015 recursive_a
JSFrame: 253.000 / 0.014 recursive_b
JSFrame: 253.000 / 0.012 recursive_a
JSFrame: 253.000 / 0.008 recursive_b
JSFrame: 253.100 / 0.200 recursive_a
JSFrame: 253.100 / 0.100 recursive_b
